requirements.txt
django
psycopg2-binary
pillow
djangorestframework
djangorestframework_simplejwt
uwsgi
raven
django-cors-headers
facebook-sdk
google-auth
django-prometheus
fcm-django
celery
redis
sqlalchemy
django-ckeditor
django-filter
python-memcached

factory_boy
drf-yasg

bandit
flake8
flake8-docstrings
flake8-quotes
flake8-mypy
flake8-print
isort
pylint
pylint-django==2.0.8
pytest
pytest-django
pytest-cov
pytest-mock